Fixes inability of data retrieval scripts to run
After the switchover to 'golden' all data retrieval scripts ceased working because the
new main.sh file didn't include relative paths (or a continuation chain.
Or a path at all, in one case).

This fixes that, and hopefully it won't occur again. Distinctly from this we
need to backport the data.

Bug: T111714
Change-Id: Iac6f784727e2f7b24098b7277c4a637f9f47990b