Finish the automated backfilling script - hopefully!
This:

1. Moves the conditional_write calls to the end of each script so
that if things are gonna break, they do it before anything gets written.
2. Finishes the automated backfilling script.

So here's how that script'll work: if you don't give it a range of dates,
it just gets yesterday's data. If you do get it a range of dates, it:

1. Gets a list of all the R scripts retrieving data, and their actual source
code.
2. Looks at each output file and sees if it's missing days covered by the
range of dates.
3. If it's not, return TRUE. We're good.
4. If it is, grep through the source text, find the retrieval script that is
erroring, source it, and run it for the date range that is missing for that
file.
5. Repeat.

You might think 'wait, but what about multiple writes per file!' and, well,
because it's checking the data file and then running the source file, both one
by one, we shouldn't end up ever running the same file twice (unless it's
the file with the error in it, which run.Rout will report if so)

Bug: T114919
Change-Id: I371ef35580b441e37a6ce78d70cf1f4ffcb9638a