Use a single date for Sys.Date()-1
This passes Sys.Date()-1 into main() from run(). The idea
is very simple; at the moment we're launching the cron job
relatively late in the day to cover for the possibility
of EventLogging replication lag. Each main() function
calculates Sys.Date()-1 on its own. When combined, this
raises the improbable but possible scenario where some
of the code takes long enough to run that the Sys.Date changes,
and some scripts run for day N and others for N+1.

Instead, with this change we'll be passing N through consistently,
for all files, and still retain the ability to run the main()
functions as self-contained blocks. Everyone wins.

Change-Id: Ia4b98b72f115128c435ec7da6610527b869fc91e