main.sh: add full paths to commands
What:
Add full command paths to the main.sh script.

Why:
This script will be called by the ExecStart of a Systemd unit. To avoid
adding an explicit WorkingPath (that is not supported in puppet for the moment)
it seems easier to just have full command paths in main.sh.

Change-Id: I4f0cac49888ab1665bcf3211e6628bd6a4f00e5e